+++ src/lib/schematics/ng-add/index.ts
@@ -6,9 +6,9 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Rule, Tree, SchematicContext, TaskId} from '@angular-devkit/schematics';
+import {Rule, SchematicContext, Tree} from '@angular-devkit/schematics';
 import {NodePackageInstallTask, RunSchematicTask} from '@angular-devkit/schematics/tasks';
-import {addPackageToPackageJson, getPackageVersionFromPackageJson} from './package-json';
+import {addPackageToPackageJson, getPackageVersionFromPackageJson} from './package-config';
 import {Schema} from './schema';
 import {hammerjsVersion, materialVersion, requiredAngularVersionRange} from './version-names';
 
@@ -21,31 +21,23 @@ import {hammerjsVersion, materialVersion, requiredAngularVersionRange} from './v
  */
 export default function(options: Schema): Rule {
   return (host: Tree, context: SchematicContext) => {
+    // Version tag of the `@angular/core` dependency that has been loaded from the `package.json`
+    // of the CLI project. This tag should be preferred because all Angular dependencies should
+    // have the same version tag if possible.
+    const ngCoreVersionTag = getPackageVersionFromPackageJson(host, '@angular/core');
+
+    addPackageToPackageJson(host, '@angular/cdk', `^${materialVersion}`);
+    addPackageToPackageJson(host, '@angular/material', `^${materialVersion}`);
+    addPackageToPackageJson(host, '@angular/animations',
+        ngCoreVersionTag || requiredAngularVersionRange);
+
+    if (options.gestures) {
+      addPackageToPackageJson(host, 'hammerjs', hammerjsVersion);
+    }
+
     // Since the Angular Material schematics depend on the schematic utility functions from the
     // CDK, we need to install the CDK before loading the schematic files that import from the CDK.
-    let installTaskId: TaskId;
-
-    if (!options.skipPackageJson) {
-      // Version tag of the `@angular/core` dependency that has been loaded from the `package.json`
-      // of the CLI project. This tag should be preferred because all Angular dependencies should
-      // have the same version tag if possible.
-      const ngCoreVersionTag = getPackageVersionFromPackageJson(host, '@angular/core');
-
-      addPackageToPackageJson(host, '@angular/cdk', `^${materialVersion}`);
-      addPackageToPackageJson(host, '@angular/material', `^${materialVersion}`);
-      addPackageToPackageJson(host, '@angular/animations',
-          ngCoreVersionTag || requiredAngularVersionRange);
-
-      if (options.gestures) {
-        addPackageToPackageJson(host, 'hammerjs', hammerjsVersion);
-      }
-
-      installTaskId = context.addTask(new NodePackageInstallTask());
-    } else {
-      installTaskId = context.addTask(new NodePackageInstallTask({
-        packageName: `@angular/cdk@^${materialVersion}`
-      }));
-    }
+    const installTaskId = context.addTask(new NodePackageInstallTask());
 
     context.addTask(new RunSchematicTask('ng-add-setup-project', options), [installTaskId]);
   };
