+++ src/lib/schematics/ng-add/package-config.ts
