+++ src/lib/schematics/ng-add/schema.json
@@ -11,11 +11,6 @@
         "$source": "projectName"
       }
     },
-    "skipPackageJson": {
-      "type": "boolean",
-      "default": false,
-      "description": "Do not add materials dependencies to package.json (e.g., --skipPackageJson)"
-    },
     "theme": {
       "enum": ["indigo-pink", "deeppurple-amber", "pink-bluegrey", "purple-green", "custom"],
       "default": "indigo-pink",
