+++ src/lib/schematics/ng-add/schema.ts
@@ -11,9 +11,6 @@ export interface Schema {
   /** Name of the project to target. */
   project: string;
 
-  /** Whether to skip package.json install. */
-  skipPackageJson: boolean;
-
   /** Whether gesture support should be set up or not. */
   gestures: boolean;
